+++ unauthenticated/js/init.js
@@ -1,5 +1,5 @@
 /*!
- * Authentic Theme 17.53 (https://github.com/qooob/authentic-theme)
+ * Authentic Theme 17.54 (https://github.com/qooob/authentic-theme)
  * Copyright 2016 Ilia Rostovtsev <programming@rostovtsev.ru>
  * Licensed under MIT (https://github.com/qooob/authentic-theme/blob/master/LICENSE)
  */
@@ -30,6 +30,7 @@ typeof settings_notification_slider_fixed == "undefined" ? settings_notification
 typeof settings_thirdparty_filemin_autofoldersize == "undefined" ? settings_thirdparty_filemin_autofoldersize = false : false;
 typeof settings_thirdparty_filemin_hide_actions == "undefined" ? settings_thirdparty_filemin_hide_actions = true : false;
 typeof settings_thirdparty_filemin_hide_toolbar == "undefined" ? settings_thirdparty_filemin_hide_toolbar = false : false;
+typeof settings_thirdparty_filemin_hovered_toolbar == "undefined" ? settings_thirdparty_filemin_hovered_toolbar = false : false;
 typeof settings_hotkeys_active == "undefined" ? settings_hotkeys_active = true : false;
 typeof settings_hotkey_toggle_modifier == "undefined" ? settings_hotkey_toggle_modifier = "altKey" : false;
 typeof settings_hotkey_toggle_key_webmin == "undefined" ? settings_hotkey_toggle_key_webmin = "w" : false;
@@ -234,7 +235,7 @@ function ___csf() {
 		k = t.contentDocument.getElementsByTagName("head")[0],
 		q = t.contentDocument.createElement("script");
 	q.type = "text/javascript";
-	q.src = "/unauthenticated/js/package.min.js?1753";
+	q.src = "/unauthenticated/js/package.min.js?1754";
 	k.appendChild(q);
 	if (o.$('iframe[name="page"]').contents().find("body.csf").length === 0) {
 		t.contentWindow.onbeforeunload = function (a) {
@@ -243,7 +244,7 @@ function ___csf() {
 		o.$('a[href="csf/"]').parent("li").addClass("sub_active").append('<span class="current"></span>').parent("ul.sub").show().prev("li").addClass("active");
 		$csf = o.$('iframe[name="page"]').contents();
 		$csf.find("html").attr("data-background-style", o.$("html").attr("data-background-style"));
-		$csf.find("head").append('				<link rel="shortcut icon" href="' + $_____link_full + '/images/favicon-webmin.ico">				<meta name="viewport" content="width=device-width, initial-scale=1.0">				<link href="' + $_____link_full + '/unauthenticated/css/package.min.css?1753" rel="stylesheet" type="text/css">			');
+		$csf.find("head").append('				<link rel="shortcut icon" href="' + $_____link_full + '/images/favicon-webmin.ico">				<meta name="viewport" content="width=device-width, initial-scale=1.0">				<link href="' + $_____link_full + '/unauthenticated/css/package.min.css?1754" rel="stylesheet" type="text/css">			');
 		$.each(o.$('link[href*="/styles.css"]'), function () {
 			if ($(this)) {
 				$csf.find("head").append('<link href="' + $_____link_full + '/unauthenticated/css/styles.css" rel="stylesheet" type="text/css">')
@@ -565,7 +566,7 @@ t__wi_p.$('iframe[name="page"]').on("load", function () {
 			return
 		}
 		__lre();
-		s("/unauthenticated/js/authentic.min.js?1753")
+		s("/unauthenticated/js/authentic.min.js?1754")
 	}
 	if (settings_loader_top && t__wi_p.t___p__xhr_l === 0) {
 		t__wi_p.NProgress.done()